% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rovr.R
\name{print.rovr}
\alias{print.rovr}
\title{Get information about Regression One-Vs-Rest object}
\usage{
\method{print}{rovr}(x, ...)
}
\arguments{
\item{x}{An object of class "rovr".}

\item{...}{Extra arguments (not used).}
}
\description{
Prints basic information about a `rovr` object
(Number of classes, regressor class).
}
