% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wap.R
\name{print.wap}
\alias{print.wap}
\title{Get information about Weighted All-Pairs object}
\usage{
\method{print}{wap}(x, ...)
}
\arguments{
\item{x}{An object of class "wap".}

\item{...}{Extra arguments (not used).}
}
\description{
Prints basic information about a `wap` object
(Number of classes and classifiers, classifier class).
}
