% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wovr.R
\name{print.wovr}
\alias{print.wovr}
\title{Get information about Weighted One-Vs-Rest object}
\usage{
\method{print}{wovr}(x, ...)
}
\arguments{
\item{x}{An object of class "wovr".}

\item{...}{Extra arguments (not used).}
}
\description{
Prints basic information about a `wovr` object
(Number of classes, classifier class).
}
