% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costprop.R
\name{summary.costprop}
\alias{summary.costprop}
\title{Get information about Cost-Proportionate classifier object}
\usage{
\method{summary}{costprop}(object, ...)
}
\arguments{
\item{object}{An object of class "costprop".}

\item{...}{Extra arguments (not used).}
}
\description{
Prints basic information about a `costprop` object
(Number of samples, rejection constant, random seed, class names, classifier class).
Same as function `print`.
}
