% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wap.R
\name{summary.wap}
\alias{summary.wap}
\title{Get information about Weighted All-Pairs object}
\usage{
\method{summary}{wap}(object, ...)
}
\arguments{
\item{object}{An object of class "wap".}

\item{...}{Extra arguments (not used).}
}
\description{
Prints basic information about a `wap` object
(Number of classes and classifiers, classifier class). Same as function `print`.
}
