% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wovr.R
\name{summary.wovr}
\alias{summary.wovr}
\title{Get information about Weighted One-Vs-Rest object}
\usage{
\method{summary}{wovr}(object, ...)
}
\arguments{
\item{object}{An object of class "wovr".}

\item{...}{Extra arguments (not used).}
}
\description{
Prints basic information about a `wovr` object
(Number of classes, classifier class). Same as function `print`.
}
